requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.36
alembic==1.13.1
psycopg[binary]==3.2.12
python-dotenv==1.0.1
redis==5.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
pydantic==1.10.18
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.9
